Update fetch.py
Should print mmC0v3rgt or similar as of now, which doesn't seem right so I have a couple questions for gourd. Must be connected to school wifi to work. Also, left a couple print statements for debugging, and didn't get a chance to test the METHOD function since I made it off campus.